added an examples folder
Cleans up the repo; move poly fills to the examples since users
might want to include their own.

Added base path option to the reader. reader now automaticly
adds a basepath to the images based on what folder the data.json
file is located. If images are actual urls, the base path is
not included